Move firebase admin helper into frontend
